More presentation stuff + various small changes
Better editing for presentations, better control.
Some hacky stuff to force the presenter to be in control of the sprite which is why this is a separate branch and not a full feature.

Small gui to manager since it takes up a lot of room.

Oh, and sorting of course. (This will be added to the main branch soon too)